Create README.md
THE CRICKET AS A THERMOMETER
Crickets are cold-blooded and take on the temperature of their surroundings. Back in 1897, a scientist named Amos Dolbear published an article “The Cricket as a Thermometer” that noted the correlation between the ambient temperature and the rate at which crickets chirp.

The insects’ muscles contract to produce chirping, based on chemical reactions. The warmer the temperature, the easier the cricket’s muscles activate, so the chirps increase. The cooler the temerpature, the slower the reaction rate, and the less frequent the chirps.

CRICKET TEMPERATURE FORMULA
The formula expressed in that article became known as Dolbear’s Law. It’s surprisingly simple:

To convert cricket chirps to degrees Fahrenheit:
Just count the number of chirps in 14 seconds, then add 40 to get the temperature.

The number you get will be an approximation of the outside temperature.

Example: 30 chirps + 40 = 70° F

To convert cricket chirps to degrees Celsius:
Count the number of chirps in 25 seconds, divide by 3, then add 4 to get the temperature.

Example: 48 chirps /(divided by) 3 + 4 = 20° C